Clean up & Fix CSS/HTML, add wrapper around sign-up-box
Rudimentary Styling is complete. I was unable to figure out how to create
Flexbox container in main body to where main.headers & main.text appear on
the left, with main.image on the right.

I am having the same issue with the sign-up-box, and so I created a sign-up
wrapper as a flexbox container for everything in order to have the
sign-up-box.headers and sign-up-box.text align to the left side of sign-up-
box with the sign-up-box .button on the bottom right of container.